# backend/api/src/product-service/swagger.yaml
openapi: 3.0.0
servers:
  - url: https://5oo81ag4l8.execute-api.eu-central-1.amazonaws.com/dev
components:
  schemas:
    GetProductRes:
      properties:
        statusCode:
          type: integer
          example: 200
        body:
          properties:
            count:
              type: integer
              example: 4
            description:
              type: string
              example: Short Product Description1
            id:
              type: string
              format: uuid
              example: 7567ec4b-b10c-48c5-9345-fc73c48a80aa
            price:
              type: integer
              example: 10
            title:
              type: string
              example: ProductOne
    PostProductReq:
      required:
        - count
        - description
        - price
        - title
      properties:
        count:
          type: integer
          example: 4
        description:
          type: string
          example: Short Product Description1
        price:
          type: integer
          example: 10
        title:
          type: string
          example: ProductOne
    GetProductsRes:
      properties:
        statusCode:
          type: integer
          example: 200
        body:
          type: array
          items:
            type: object
            properties:
              count:
                type: integer
                example: 4
              description:
                type: string
                example: Short Product Description1
              id:
                type: string
                format: uuid
                example: 7567ec4b-b10c-48c5-9345-fc73c48a80aa
              price:
                type: integer
                example: 10
              title:
                type: string
                example: ProductOne
    NotFoundProductRes:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        body:
          type: string
          example: Product not found
  securitySchemes: {}
info:
  title: Product-service
  description: Product-service OpenAPI
  version: '2'
paths:
  /products:
    get:
      operationId: getProductsList
      summary: Get all products
      description: Get all products
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProductsRes'
    post:
      operationId: postProduct
      summary: Creating a new item in a products table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostProductReq'
      responses:
        '201':
          description: Status 201 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProductRes'
        '400':
          description: Status 400 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundProductRes'
  /products/{productId}:
    get:
      operationId: getProductsById
      summary: Get product by id
      description: Get product by id
      parameters:
        - in: path
          name: productId
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProductRes'
        '404':
          description: Status 404 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundProductRes'